+++ docker-with-goland/main.tf
@@ -1,0 +1,159 @@
+terraform {
+  required_providers {
+    coder = {
+      source  = "coder/coder"
+      version = "~> 0.5.3"
+    }
+    docker = {
+      source  = "kreuzwerker/docker"
+      version = "~> 2.22.0"
+    }
+  }
+}
+
+variable "dotfiles_uri" {
+  description = <<-EOF
+  Dotfiles repo URI (optional)
+
+  see https://dotfiles.github.io
+  EOF
+  default = "git@github.com:sharkymark/dotfiles.git"
+}
+
+variable "image" {
+  description = "Container Image"
+  default     = "codercom/enterprise-goland:ubuntu"
+  validation {
+    condition = contains([
+      "codercom/enterprise-goland:ubuntu"
+    ], var.image)
+    error_message = "Invalid container image!"  
+}
+}
+
+provider "docker" {
+  host = "unix:///var/run/docker.sock"
+}
+
+provider "coder" {
+}
+
+data "coder_workspace" "me" {
+}
+
+
+resource "coder_agent" "coder" {
+  os   = "linux"
+  arch = "amd64"
+  dir = "/home/coder"
+  startup_script = <<EOT
+#!/bin/bash
+
+# change owner to coder so jetbrains projector CLI can create a config
+sudo chown -R coder:coder /opt &
+
+# use coder CLI to clone and install dotfiles
+coder dotfiles -y ${var.dotfiles_uri}
+
+# install and start code-server
+curl -fsSL https://code-server.dev/install.sh | sh
+code-server --auth none --port 13337 &
+
+# install jetbrains projector CLI and create goland config
+pip3 install projector-installer --user
+/home/coder/.local/bin/projector --accept-license 
+
+/home/coder/.local/bin/projector config add goland /opt/goland --force --use-separate-config --port 9001 --hostname localhost
+/home/coder/.local/bin/projector run goland &
+
+# clone 3 Go repos
+mkdir -p ~/.ssh
+ssh-keyscan -t ed25519 github.com >> ~/.ssh/known_hosts
+git clone --progress git@github.com:sharkymark/chi_helloworld.git &
+git clone --progress git@github.com:sharkymark/commissions.git &
+git clone --progress git@github.com:coder/coder.git &
+
+
+# install VS Code extensions into code-server
+SERVICE_URL=https://open-vsx.org/vscode/gallery ITEM_URL=https://open-vsx.org/vscode/item code-server --install-extension golang.go
+
+EOT
+}
+
+# code-server
+resource "coder_app" "code-server" {
+  agent_id      = coder_agent.coder.id
+  name          = "VS Code"
+  icon          = "/icon/code.svg"
+  url           = "http://localhost:13337?folder=/home/coder"
+  subdomain = false
+  share     = "owner"
+
+  healthcheck {
+    url       = "http://localhost:13337/healthz"
+    interval  = 5
+    threshold = 15
+  } 
+}
+
+resource "coder_app" "goland" {
+  agent_id      = coder_agent.coder.id
+  name          = "GoLand"
+  icon          = "/icon/goland.svg"
+  url           = "http://localhost:9001/"
+  subdomain = false
+  share     = "owner"
+
+  healthcheck {
+    url       = "http://localhost:9001/healthz"
+    interval  = 10
+    threshold = 20
+  } 
+}
+
+resource "docker_container" "workspace" {
+  count = data.coder_workspace.me.start_count
+  image = "${var.image}"
+  # Uses lower() to avoid Docker restriction on container names.
+  name     = "coder-${data.coder_workspace.me.owner}-${lower(data.coder_workspace.me.name)}"
+  hostname = lower(data.coder_workspace.me.name)
+  dns      = ["1.1.1.1"]
+
+
+  # Use the docker gateway if the access URL is 127.0.0.1
+  #entrypoint = ["sh", "-c", replace(coder_agent.dev.init_script, "127.0.0.1", "host.docker.internal")]
+
+  # Use the docker gateway if the access URL is 127.0.0.1
+  command = [
+    "sh", "-c",
+    <<EOT
+    trap '[ $? -ne 0 ] && echo === Agent script exited with non-zero code. Sleeping infinitely to preserve logs... && sleep infinity' EXIT
+    ${replace(coder_agent.coder.init_script, "/localhost|127\\.0\\.0\\.1/", "host.docker.internal")}
+    EOT
+  ]
+
+
+  env        = ["CODER_AGENT_TOKEN=${coder_agent.coder.token}"]
+  volumes {
+    container_path = "/home/coder/"
+    volume_name    = docker_volume.coder_volume.name
+    read_only      = false
+  }  
+  host {
+    host = "host.docker.internal"
+    ip   = "host-gateway"
+  }
+}
+
+resource "docker_volume" "coder_volume" {
+  name = "coder-${data.coder_workspace.me.owner}-${data.coder_workspace.me.name}"
+}
+
+resource "coder_metadata" "workspace_info" {
+  count       = data.coder_workspace.me.start_count
+  resource_id = docker_container.workspace[0].id   
+  item {
+    key   = "image"
+    value = "${var.image}"
+  }     
+}